Moved base_agent.py to common folder
